+++ src/test/utility_test.rs
@@ -1,4 +1,7 @@
+use crate::utility::*;
+
 mod kernel_pca_test;
+mod label_encoding_test;
 mod linear_discriminant_analysis_test;
 mod principal_component_analysis_test;
 mod t_sne_test;